fix: address pr comments
