+++ libs/application/templates/national-agency-for-children-and-families/child-protection-notification/src/utils/constants.ts
@@ -19,7 +19,7 @@ export enum ApiModuleActions {
   getCategories = 'getCategories',
   getProtectiveFactors = 'getProtectiveFactors',
   getGenders = 'getGenders',
-  getChildInformation = 'getChildInformation',
+  getLanguageEnvironments = 'getLanguageEnvironments',
   getUrgencyAssessments = 'getUrgencyAssessments',
   getPostalCodes = 'getPostalCodes',
   getPronouns = 'getPronouns',
